# swagger.yml
openapi: 3.0.3
info:
  title: Soda Soft Drink Store API
  version: 1.0.0
  description: >
    Codecademy E-Commerce REST API for the Soda Soft Drink Store.
    Register/login to receive a JWT, browse soda products, manage a cart,
    and place orders.
servers:
  - url: http://localhost:4001/api
    description: Local development

tags:
  - name: Auth
    description: Registration and login
  - name: Products
    description: Browse and search soda drinks
  - name: Cart
    description: Manage the authenticated user's cart
  - name: Orders
    description: Checkout and order history
  - name: Users
    description: Account profile

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: renee@example.com }
        password: { type: string, example: hunter2 }
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: Renee }
        email: { type: string, example: renee@example.com }
        password: { type: string, example: hunter2 }
    AuthResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        token: { type: string }
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string }
        name: { type: string }
        created_at: { type: string, format: date-time }
    Product:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: Classic Cola }
        description: { type: string }
        price: { type: number, format: float, example: 1.99 }
        stock: { type: integer, example: 100 }
        flavor: { type: string, example: cola }
        size: { type: string, example: 330ml }
        image_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    Cart:
      type: object
      properties:
        id: { type: integer }
        total: { type: number, format: float }
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItem' }
    CartItem:
      type: object
      properties:
        id: { type: integer }
        product_id: { type: integer }
        name: { type: string }
        price: { type: number, format: float }
        quantity: { type: integer }
        line_total: { type: number, format: float }
    Order:
      type: object
      properties:
        id: { type: integer }
        total: { type: number, format: float }
        status: { type: string, example: pending }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
    OrderItem:
      type: object
      properties:
        product_id: { type: integer }
        name: { type: string }
        quantity: { type: integer }
        price: { type: number, format: float }
        line_total: { type: number, format: float }
    Error:
      type: object
      properties:
        error: { type: string }

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '409':
          description: Email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Credentials' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /products:
    get:
      tags: [Products]
      summary: List / search soda products
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Match name or description
        - in: query
          name: flavor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [Products]
      summary: Create a product (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }

  /products/{id}:
    get:
      tags: [Products]
      summary: Get single drink details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Products]
      summary: Update a product (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
    delete:
      tags: [Products]
      summary: Delete a product (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted

  /cart:
    get:
      tags: [Cart]
      summary: View the current cart
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
    delete:
      tags: [Cart]
      summary: Empty the cart
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The emptied cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }

  /cart/items:
    post:
      tags: [Cart]
      summary: Add a product to the cart
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: integer, example: 1 }
                quantity: { type: integer, example: 2 }
      responses:
        '201':
          description: Updated cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }

  /cart/items/{productId}:
    put:
      tags: [Cart]
      summary: Set a line's quantity (0 removes it)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity: { type: integer, example: 3 }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
    delete:
      tags: [Cart]
      summary: Remove a line from the cart
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }

  /orders:
    post:
      tags: [Orders]
      summary: Checkout — create an order from the cart
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: The created order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Empty cart
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [Orders]
      summary: List the current user's orders
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }

  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get a single order with items
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /users/me:
    get:
      tags: [Users]
      summary: Get the authenticated user's profile
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    put:
      tags: [Users]
      summary: Update name / email
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }